+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## Unreleased
 
+## 5.16.0 - 2019-09-29
+
+### Enhancements
+
+- Changed user interaction when asking user for telemetry permission
+- Disabling .net optimization scheduled tasks on all 2012R2+ machines
+- Updated Sql2017ManagementStudio link to version 18.2
+
+### Bug fixes
+- Azure module test method updated to actually locate the Az module (Fixed #671)
+
 ## 5.15.0 - 2019-09-20
 
 ### Enhancements